+++ src/custom-types.ts
@@ -198,7 +198,8 @@ export const ElysiaType = {
 		Type.Transform(
 			Type.Union([
 				Type.String({
-					format: 'ObjectString'
+					format: 'ObjectString',
+					default: ''
 				}),
 				Type.Object(properties, options)
 			])
@@ -215,7 +216,7 @@ export const ElysiaType = {
 			})
 			.Encode((value) => JSON.stringify(value)) as any as TObject<T>,
 	File: TypeSystem.Type<File, ElysiaTypeOptions.File>('File', validateFile),
-	Files: (options: ElysiaTypeOptions.Files) =>
+	Files: (options: ElysiaTypeOptions.Files = {}) =>
 		Type.Transform(Type.Union([Files(options)]))
 			.Decode((value) => {
 				if (Array.isArray(value)) return value